add game saving, loading, undo, liberation
Game loading is done using a sort of recursive descent parser. Not
muche error checking is done.
Captured pieces now go to the jail nearest to their color. As such,
jails are now more like hospitals.
For the record, game loading was implemented at first for faster
testing of liberation, which takes a minimum of ~30 steps to
trigger, but it ended up being lots of fun.